+++ brewtils/decorators.py
@@ -469,6 +469,24 @@ def shutdown(_wrapped=None):
     return _wrapped
 
 
+def startup(_wrapped=None):
+    """Decorator for specifying a function to run before a plugin is running.
+
+    for example::
+
+        @startup
+        def pre_running(self):
+            # Run pre-running processing
+            return
+
+    Args:
+        _wrapped: The function to decorate. This is handled as a positional argument and
+            shouldn't be explicitly set.
+    """
+    _wrapped._startup = True
+    return _wrapped
+
+
 def subscribe(_wrapped=None, topic: str = None, topics=[]):
     """Decorator for specifiying topic to listen to.
 
@@ -527,6 +545,24 @@ def _parse_shutdown_functions(client):
             shutdown_functions.append(method)
 
     return shutdown_functions
+
+
+def _parse_startup_functions(client):
+    # type: (object) -> List[Callable]
+    """Get a list of callable fields labeled with the startup annotation
+
+    This will iterate over everything returned from dir, looking for metadata added
+    by the startup decorator.
+    """
+
+    startup_functions = []
+
+    for attr in dir(client):
+        method = getattr(client, attr)
+        if callable(method) and getattr(method, "_startup", False):
+            startup_functions.append(method)
+
+    return startup_functions
 
 
 def _parse_client(client):