+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ TBD
 
 - Added new annotation/configuration support for shutdown functions. These functions will be executed at the start
   of the shutdown process.
+- Added new annotation/configuration support for startup functions. These functions will be executed after `Plugin().run()` 
+  has completed startup processes
+
 
 3.28.0
 ------